+++ src/components/Spinner/Spinner.js
@@ -1,30 +1,39 @@
 import React, { memo } from 'react';
 import PropTypes from 'prop-types';
 
-import { Pane } from '@folio/stripes/components';
+import {
+  Pane,
+  PaneHeader,
+} from '@folio/stripes/components';
 
 import { Preloader } from '../Preloader';
 
 export const Spinner = memo(({
-  entity,
   entity: {
     props: {
       onClose,
       paneId,
     },
   },
-}) => (
-  <Pane
-    id={paneId}
-    defaultWidth="fill"
-    fluidContentWidth
-    paneTitle=""
-    dismissible
-    lastMenu={entity.renderLastMenu()}
-    onClose={onClose}
-  >
-    <Preloader />
-  </Pane>
-));
+}) => {
+  const header = renderProps => (
+    <PaneHeader
+      {...renderProps}
+      id={paneId}
+      dismissible
+      onClose={onClose}
+    />
+  );
+
+  return (
+    <Pane
+      defaultWidth="fill"
+      fluidContentWidth
+      renderHeader={header}
+    >
+      <Preloader />
+    </Pane>
+  );
+});
 
 Spinner.propTypes = { entity: PropTypes.object.isRequired };