+++ src/settings/MatchProfiles/ViewMatchProfile.js
@@ -20,6 +20,7 @@ import {
   Accordion,
   AccordionSet,
   ConfirmationModal,
+  PaneHeader,
 } from '@folio/stripes/components';
 import {
   ViewMetaData,
@@ -27,12 +28,8 @@ import {
   TagsAccordion,
 } from '@folio/stripes/smart-components';
 
+import { getFieldMatched } from '../../utils';
 import {
-  createLayerURL,
-  getFieldMatched,
-} from '../../utils';
-import {
-  LAYER_TYPES,
   ENTITY_KEYS,
   SYSTEM_USER_ID,
   SYSTEM_USER_NAME,
@@ -49,8 +46,10 @@ import {
   FlexibleForm,
   ProfileAssociator,
 } from '../../components';
-import { FOLIO_RECORD_TYPES } from '../../components/ListTemplate';
-import { LastMenu } from '../../components/ActionMenu/ItemTemplates/LastMenu';
+import {
+  FOLIO_RECORD_TYPES,
+  INCOMING_RECORD_TYPES,
+} from '../../components/ListTemplate';
 
 import sharedCss from '../../shared.css';
 import styles from './ViewMatchProfile.css';
@@ -78,13 +77,6 @@ export class ViewMatchProfile extends Component {
         records: PropTypes.arrayOf(PropTypes.object),
       }),
     }).isRequired,
-    location: PropTypes.oneOfType([
-      PropTypes.shape({
-        search: PropTypes.string.isRequired,
-        pathname: PropTypes.string.isRequired,
-      }).isRequired,
-      PropTypes.string.isRequired,
-    ]),
     match: PropTypes.shape({ params: PropTypes.shape({ id: PropTypes.string }).isRequired }).isRequired,
     tagsEnabled: PropTypes.bool,
     onClose: PropTypes.func.isRequired,
@@ -153,15 +145,6 @@ export class ViewMatchProfile extends Component {
     />
   );
 
-  renderLastMenu = record => (
-    <LastMenu
-      caption="ui-data-import.edit"
-      location={createLayerURL(this.props.location, LAYER_TYPES.EDIT)}
-      style={{ visibility: !record ? 'hidden' : 'visible' }}
-      dataAttributes={{ 'data-test-edit-item-button': '' }}
-    />
-  );
-
   getValue = (fields, label) => {
     const field = fields.find(item => item.label === label);
 
@@ -174,12 +157,39 @@ export class ViewMatchProfile extends Component {
     return !isEmpty(element) ? <FormattedMessage id={element.label} /> : undefined;
   };
 
-  render() {
+  renderPaneHeader = renderProps => {
     const {
       onClose,
       paneId,
-      tagsEnabled,
     } = this.props;
+
+    const { record: matchProfile } = this.matchProfileData;
+
+    const paneTitle = (
+      <AppIcon
+        size="small"
+        app="data-import"
+        iconKey="matchProfiles"
+      >
+        {matchProfile.name}
+      </AppIcon>
+    );
+
+    return (
+      <PaneHeader
+        {...renderProps}
+        id={paneId}
+        paneTitle={paneTitle}
+        paneSub={<FormattedMessage id="ui-data-import.matchProfileName" />}
+        actionMenu={this.renderActionMenu}
+        dismissible
+        onClose={onClose}
+      />
+    );
+  };
+
+  render() {
+    const { tagsEnabled } = this.props;
     const { showDeleteConfirmation } = this.state;
 
     const {
@@ -191,16 +201,6 @@ export class ViewMatchProfile extends Component {
     if (!matchProfile || !hasLoaded) {
       return <Spinner entity={this} />;
     }
-
-    const paneTitle = (
-      <AppIcon
-        size="small"
-        app="data-import"
-        iconKey="matchProfiles"
-      >
-        {matchProfile.name}
-      </AppIcon>
-    );
 
     // MatchProfiles sample data does not contain user Ids because of back-end limitations
     // and therefore it is required to add it manually on UI side
@@ -273,12 +273,34 @@ export class ViewMatchProfile extends Component {
         existingRecordType: get(matchDetails, 'existingRecordType', ''),
       }],
     };
-    const existingRecordLabel = <FormattedMessage id={FOLIO_RECORD_TYPES[matchProfile.existingRecordType].captionId} />;
+    const existingRecordLabel = FOLIO_RECORD_TYPES[matchProfile.existingRecordType]
+      ? <FormattedMessage id={FOLIO_RECORD_TYPES[matchProfile.existingRecordType].captionId} />
+      : '';
+    const incomingRecordLabel = INCOMING_RECORD_TYPES[matchProfile.incomingRecordType]
+      ? <FormattedMessage id={INCOMING_RECORD_TYPES[matchProfile.incomingRecordType].captionId} />
+      : '';
     const componentsProps = {
       'panel-existing': {
         id: 'panel-existing-view',
         existingRecordType: matchProfile.existingRecordType,
+        incomingRecordType: matchProfile.incomingRecordType,
         isEditable: false,
+      },
+      'incoming-record-section': {
+        label: (
+          <FormattedMessage
+            id="ui-data-import.match.incoming.record"
+            values={{ recordType: incomingRecordLabel }}
+          />
+        ),
+      },
+      'incoming-record-field': {
+        label: (
+          <FormattedMessage
+            id="ui-data-import.match.incoming.record.field"
+            values={{ recordType: incomingRecordLabel }}
+          />
+        ),
       },
       'existing-record-section': {
         label: (
@@ -300,15 +322,9 @@ export class ViewMatchProfile extends Component {
 
     return (
       <Pane
-        id={paneId}
+        renderHeader={this.renderPaneHeader}
         defaultWidth="620px"
         fluidContentWidth
-        paneTitle={paneTitle}
-        paneSub={<FormattedMessage id="ui-data-import.matchProfileName" />}
-        actionMenu={this.renderActionMenu}
-        lastMenu={this.renderLastMenu(matchProfile)}
-        dismissible
-        onClose={onClose}
       >
         <TitleManager record={matchProfile.name} />
         <Headline